Move some test utilities into the `vdaf_poc` package
Pull `test_flp_bbcggi19` and `test_vdaf` into into the installable
package so they can be used elsewhere.
