+++ poc/tests/test_flp_bbcggi19.py
@@ -1,12 +1,12 @@
 import unittest
 from typing import TypeVar
 
-from vdaf_poc.common import next_power_of_2
-from vdaf_poc.field import FftField, Field64, Field96, Field128, poly_eval
-from vdaf_poc.flp import Flp, run_flp
-from vdaf_poc.flp_bbcggi19 import (Count, FlpBBCGGI19, Gadget, Histogram, Mul,
+from vdaf_poc.field import FftField, Field64, Field96, Field128
+from vdaf_poc.flp import Flp
+from vdaf_poc.flp_bbcggi19 import (Count, FlpBBCGGI19, Histogram, Mul,
                                    MultihotCountVec, PolyEval, Range2, Sum,
-                                   SumOfRangeCheckedInputs, SumVec, Valid)
+                                   SumOfRangeCheckedInputs, SumVec, Valid,
+                                   test_flp_bbcggi19, test_gadget)
 
 Measurement = TypeVar("Measurement")
 AggResult = TypeVar("AggResult")
@@ -43,53 +43,6 @@ class TestMultiGadget(Valid[int, int, Field64]):
 
     def decode(self, output: list[Field64], _num_measurements: int) -> int:
         return output[0].as_unsigned()
-
-
-def test_gadget(g: Gadget, field: type[F], test_length: int) -> None:
-    """
-    Test for equivalence of `Gadget.eval()` and `Gadget.eval_poly()`.
-    """
-    meas_poly = []
-    meas = []
-    eval_at = field.rand_vec(1)[0]
-    for _ in range(g.ARITY):
-        meas_poly.append(field.rand_vec(test_length))
-        meas.append(poly_eval(field, meas_poly[-1], eval_at))
-    out_poly = g.eval_poly(field, meas_poly)
-
-    want = g.eval(field, meas)
-    got = poly_eval(field, out_poly, eval_at)
-    assert got == want
-
-
-def test_flp_bbcggi19(
-        flp: FlpBBCGGI19[Measurement, AggResult, F],
-        test_cases: list[tuple[list[F], bool]]) -> None:
-    for (g, g_calls) in zip(flp.valid.GADGETS, flp.valid.GADGET_CALLS):
-        test_gadget(g, flp.field, next_power_of_2(g_calls + 1))
-
-    for (i, (meas, expected_decision)) in enumerate(test_cases):
-        assert len(meas) == flp.MEAS_LEN
-        assert len(flp.truncate(meas)) == flp.OUTPUT_LEN
-
-        # Evaluate validity circuit.
-        joint_rand = flp.field.rand_vec(flp.JOINT_RAND_LEN)
-        v = flp.valid.eval(meas, joint_rand, 1)
-        if (v == [flp.field(0)] * flp.valid.EVAL_OUTPUT_LEN) != \
-                expected_decision:
-            print('{}: test {} failed: validity circuit returned {}'.format(
-                flp.valid.__class__.__name__, i, v))
-
-        # Run the FLP.
-        decision = run_flp(flp, meas, 2)
-        if decision != expected_decision:
-            print(
-                '{}: test {} failed: proof evaluation resulted in {}; want {}'
-                .format(
-                    flp.valid.__class__.__name__, i, decision,
-                    expected_decision,
-                )
-            )
 
 
 class TestAverage(Sum):
