+++ poc/tests/test_vdaf_prio3.py
@@ -3,10 +3,10 @@ from typing import TypeVar
 
 from tests.test_flp import FlpTest
 from tests.test_flp_bbcggi19 import TestAverage
-from tests.vdaf_util import test_vdaf
 from vdaf_poc.common import TEST_VECTOR
 from vdaf_poc.field import FftField, Field64, Field128
 from vdaf_poc.flp_bbcggi19 import FlpBBCGGI19
+from vdaf_poc.vdaf import test_vdaf
 from vdaf_poc.vdaf_prio3 import (Prio3, Prio3Count, Prio3Histogram,
                                  Prio3MultihotCountVec, Prio3Sum, Prio3SumVec,
                                  Prio3SumVecWithMultiproof)
